make panel actually use its constructor parameters

+++ EvoSim/Renderer/Headers/Panel.h
@@ -6,7 +6,7 @@
 
 class Panel : public UiElement {
 public:
-	Panel(int x, int y, int height, int width, Camera* camera, glm::vec3 color = glm::vec3(1.0f, 1.0f, 1.0f));
+	Panel(float x, float y, float height, float width, Camera* camera, glm::vec3 color = glm::vec3(1.0f, 1.0f, 1.0f));
 	void onClick() override;
 	void insertSprites(Sprite** spriteArray) const override;
 	unsigned int getNumSprites() const override;
